fix signature for 6.11

+++ XIVAuras/Helpers/SpellHelpers.cs
@@ -139,7 +139,7 @@ namespace XIVAuras.Helpers
             return;
         }
 
-        public unsafe bool CanUseAction(uint actionId, ActionType type = ActionType.Spell, uint targetId = 0xE000_0000)
+        public unsafe bool CanUseAction(uint actionId, ActionType type = ActionType.Spell, long targetId = 0xE000_0000)
         {
             return _actionManager->GetActionStatus(type, actionId, targetId, 0, 1) == 0;
         }
